#408: Moved PassiveEquipmentCategoryKillerDamageIncreaseParser into the model folder
The file was not changed yet.
