+++ src/app/ffbe/mappers/effects/passive-effect-parser.factory.ts
@@ -16,7 +16,7 @@ import {PassiveDebuffsResistanceParser} from './passives/passive-debuffs-resista
 import {PassiveEsperGroupSummonParser} from './passives/passive-esper-group-summon.parser';
 import {PassiveJumpDamageIncreaseParser} from './passives/passive-jump-damage-increase.parser';
 import {PassiveSkillModifierIncreaseParser} from './passives/passive-skill-modifier-increase.parser';
-import {PassiveEquipmentCategoryKillerDamageIncreaseParser} from './passives/passive-equipment-category-killer-damage-increase.parser';
+import {PassiveEquipmentCategoryKillerDamageIncreaseParser} from '../../model/effects/passives/passive-equipment-category-killer-damage-increase.parser';
 import {PassiveLbUpgradeParser} from './passives/passive-lb-upgrade.parser';
 import {PassiveLbUpgradeHpThresholdParser} from './passives/passive-lb-upgrade-hp-threshold.parser';
 import {PassiveEsperStatsIncreaseParser} from './passives/passive-esper-stats-increase.parser';
