+++ src/app/ffbe/model/effects/passives/passive-equipment-category-killer-damage-increase.parser.ts
@@ -1,7 +1,7 @@
-import {EffectParser} from '../effect-parser';
-import {Skill} from '../../../model/skill.model';
+import {EffectParser} from '../../../mappers/effects/effect-parser';
+import {Skill} from '../../skill.model';
 import {FFBE_MONSTER_TYPES} from '../../../ffbe.constants';
-import {HTML_LINE_RETURN} from '../skill-effects.mapper';
+import {HTML_LINE_RETURN} from '../../../mappers/effects/skill-effects.mapper';
 
 export class PassiveEquipmentCategoryKillerDamageIncreaseParser extends EffectParser {
   public parse(effect: Array<any>, skill: Skill): string {
